añadido medida y cantidad a la ventana de receta y arreglado en el caso de que no se pueda acceder a la imagen de la receta

+++ iosAPP/BD/MeasureDataHelper.swift
@@ -16,7 +16,7 @@ class MeasureDataHelper: DataHelperProtocol {
     static let measureId = Expression<Int64>("measureId")
     static let measureIdServer = Expression<Int64>("measureIdServer")
     static let measure = Expression<String>("measure")
-    static let quantity = Expression<Int64>("quantity")
+    static let quantity = Expression<Double>("quantity")
     static let recipeId = Expression<Int64>("recipeId")
     static let ingredientId = Expression<Int64>("ingredientId")
